HSM-14-19 decomposition cont.: carve NotebookUI + widen shared formatting helpers (app 3,107 -> 2,880)
- NotebookUI.swift (227) — the PencilKit notebook UI (HSM-8-02, the magic pencil): canvas, pages,
  note cards, view-model wiring. The Notebook/NotebookStore seam stays in RuntimeCore.
- Widened 7 top-level private formatting helpers (artifactTint/artifactGlyph/profileIcon/
  profileBlurb/clockString/plainPreview/noteArtifactTitle) to internal so the carved modules share
  them — the cross-file dependency the verified build caught (NotebookUI needed artifactGlyph).

Device-arch xcodebuild BUILD SUCCEEDED. Nine modules carved; MeetingCaptureApp.swift 8,042 -> 2,880.
